+++ plutarch-orphanage/CHANGELOG.md
@@ -10,6 +10,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/).
 
 * `V3` version of `MintValue` that does not include zero Ada
 * QuickCheck type class instances for `AssetClass`
+* `UnsortedAssocMap` wrapper to `PlutusLedgerApi.V1.Orphans`
 
 ### Changed